example: provide a wider range of application
This commit intends to display a wider range of usage scenarios of the
plugin. Therefore, it:
  - introduces a media folder directly in the docs folder,
  - provides a third test.drawio diagram in the media folder and
  - creates an additional page, test/test3/index.md.
The newly introduced references the .drawio diagram in the media folder,
i.e. unlike in the two existing pages the diagram is not colocated with
the markdown file of the page.

In addition, this commit lets the page hierarchy be discovered rather than
explicitely defined to allow for proper debugging of use_directory_urls.

on-behalf-of: @PHOENIXCONTACT github@phoenixcontact.com