fix: enable correct relative links to drawio files
This commit provides a permanent fix such that files can be properly
specified with relative paths - even to different hierarchy levels.
Prior, a workaround was deployed that just stripped ../ from the start
of relative paths. This, however, had negative side effects when the
.drawio files were not colocated with the markdown page (e.g. if they
are stored in a separate media file structure):
  - paths then had to specified with an extra hierarchy layer, breaking
    IDE completions
  - even then, mkdocs very own build pipeline threw a warning in the
    build process, detailing that it cannot find the referenced file
    (because it does not strip ../ in the beginning of paths)

on-behalf-of: @PHOENIXCONTACT github@phoenixcontact.com